+++ R/transmute.R
@@ -23,6 +23,7 @@ dt_transmute.tidytable <- function(.data, ..., by = NULL) {
 
   dots <- enexprs(...)
   by <- enexpr(by)
+  by <- vec_selector_by(.data, !!by)
 
   eval_expr(
     .data[, list(!!!dots), by = !!by]