+++ tests/testthat/test-mutate_variants.R
@@ -89,6 +89,15 @@ test_that("mutate_across(): .cols works with is.numeric", {
   expect_equal(df$end_x, c(3,3,3))
 })
 
+test_that("mutate_across(): modify-by-reference doesn't occur", {
+  df <- data.table(x_start = c(1,1,1), end_x = c(2,2,2), z = c("a", "a", "b"))
+  df %>%
+    dt_mutate_across(is.numeric, function(.x) .x + 1)
+
+  expect_equal(df$x_start, c(1,1,1))
+  expect_equal(df$end_x, c(2,2,2))
+})
+
 test_that("mutate_across(): .cols works with is.numeric with data.frame", {
   df <- data.frame(x_start = c(1,1,1), end_x = c(2,2,2), z = c("a", "a", "b"))
   df <- df %>%
@@ -159,6 +168,20 @@ test_that("mutate_across() works with by", {
 
   results_df <- test_df %>%
     dt_mutate_across(c(x, y), ~ mean(.x), by = z)
+
+  expect_named(results_df, c("x", "y", "z"))
+  expect_equal(results_df$x, c(1.5, 1.5, 3))
+  expect_equal(results_df$y, c(4.5, 4.5, 6))
+})
+
+test_that("mutate_across() works with by enhanced selection", {
+  test_df <- data.table::data.table(
+    x = c(1,2,3),
+    y = c(4,5,6),
+    z = c("a","a","b"))
+
+  results_df <- test_df %>%
+    dt_mutate_across(c(x, y), ~ mean(.x), by = is.character)
 
   expect_named(results_df, c("x", "y", "z"))
   expect_equal(results_df$x, c(1.5, 1.5, 3))