+++ tests/testthat/test-pull.R
@@ -1,8 +1,8 @@
-test_that("expect error when empty", {
+test_that("pulls last when empty", {
   test_df <- data.table(x = c(1,2,3), y = c(4,5,6))
 
 
-  expect_error(dt_pull(test_df), "var must be supplied")
+  expect_equal(dt_pull(test_df), c(4,5,6))
 })
 
 test_that("returns a vec", {